Split CC/CXX environment variables by space characters
People sometimes set CC="ccache gcc" or CC="distcc gcc", but gcc-rs
currently doesn't recognize arguments in CC/CXX and treats the whole
value as the compiler executable.
With this patch, CC/CXX environment variables will be splitted by
whitespace charactors. The first string is used as the compiler, and the
rest are used as arguments for the compiler.