Add .link option to compile to simplify linking
Some helper libraries in rust require to be linked as well. For example:

rs/
   src/
       main.rs
       native.c

Where main.rs links to native, and native.c will be linked to some platform
library, such as netsnmp.

This allows build.rs to contain:

gcc::Config::new()
   .link("netsnmp")
   ....

Which hugely simplifies the build process, and is very reliable compared to
trying to set cflags.