Add weight to the gcc parallel tasks
Tasks are not parallelized until their total "cost" exceeds 10k.
The weight of a vector is the length of the vector. The weight of
the task is increased by 5% for each additional task added (eg
for_each), but for compilation phases, because the step itself is
quite expensive, it is worth increasing the weight to force maximum
parallelism.