Detect and use `sccache` via `RUSTC_WRAPPER`
If no other C/C++ caching tool is found by inspecting `CC` and `CXX`,
`RUSTC_WRAPPER` is tested to see if an output-caching wrapper for
`rustc` is in use. If that is the case and it is a wrapper known to also
support C/C++ caching, use it.

(Also correct/clarify a misnamed variable that caused me some confusion
looking over the code.)