Tell MSVC to link with debug CRT on debug builds
When doing a debug build, emits /MTd or /MDd instead of /MT or /MD,
matching the usual flags for C++ code. When linking statically, link.exe
requires all objects to be built with the same /M* flags, so hopefully
this makes it easier to consume Rust static libraries while debugging.