For clang-cl, separate flags/options from input file to avoid `-Wslash-u-filename`. (#513)
This avoids a `clang-cl` issue common when cross-compiling from macOS
to Windows: the `-Wslash-u-filename` where a `/Users/...` path is
interpreted as the `cl.exe` flag `/U`.

Not all compilers recognize `--` so this is limited to `clang-cl`.

There are no existing tests mocking out `clang-cl` and surface efforts
didn't succeed, so that will have to wait for follow-up.