Always specify architecture for x86_64 / aarch64 Darwin
This has a few benefits. The first is that it's not required to
specify `CFLAGS_aarch64_apple_darwin='-arch arm64'` when
cross-compiling from x86_64 to aarch64 (or `-arch x86_64` the other
way around).

Another is that the `cc` frontend performs some snooping of the
process hierarchy — if any parent is running in the Rosetta emulation
layer, it will default to targeting x86_64. This means that running
`cc` on a Developer Transition Kit is likely to suddenly and magically
revert to x86_64 if any piece of the chain is emulated (`rustc`,
`cargo`, `rustup`, the shell, even the terminal emulator!).

It's likely that the `-m64` option is superfluous with these new
options, but I don't see any harm in it either.