Split `s` out of the `ar` invocation
This commit splits out the `s` command from the `ar` invocation to
happen as part of a separate step after the archive is fully assembled.
Turns out #569 has an issue on Linux where `qs` implies `r` (??), so to
get the append-only behavior of `q` we need to omit the `s`.

Closes #568 (for real this time?)