Improve CUDA support (#612)
* Perform CUDA --device-link.

This allows to perform the final link with system linker.

* Add 'cudart' method mimicking the '--cudart' nvcc command-line option.

Try to locate the library in standard location relative to nvcc command.
If it fails, user is held responsible for specifying one in RUSTFLAGS.

* Add dummy CUDA test to cc-test.

Execution is bound to fail without card, but the failure is ignored.
It's rather a compile-n-link test. The test is suppressed if 'nvcc'
is not found on the $PATH.

* Add dummy CUDA CI test.

* Harmonize CUDA support with NVCC default --cudart static.

This can interfere with current deployments in the wild, in which
case some adjustments might be required. Most notably one might
have to add .cuda("none") to the corresponding Builder instantiation
to restore the original behaviour.