Clean up the GCC triple targeting code
Identify which if statements can coincide and which are mutually exclusive and
then reorganize them into `if .. else if ..` blocks to make it clearer which
code may or may not apply to any given target.

Remove duplicated handling of certain triple cases (e.g. armv7a).

Move everything that's not to do with the target architecture/instruction
set/calling convention/fpu to the top, then deal with the rest (the messiest and
most complicated bits of the code).