add a function to populate flags from an arbitrary environment variable
My specific use-case is to add ability for the end users to specify
crate-specific environment variables, but with how general this method
is I could imagine numerous other applications too.

Why does this need to be in `cc-rs`, you ask? The code to split up the
`CFLAGS` or `CXXFLAGS` environment variables into flags seems trivial
today, but I worry that it might not remain the case indefinitely into
the future. If everybody is expected to do this in their own build
scripts, we’re certain to run into a very inconsistent experience as
soon as parsing needs an adjustment.