added a function to remove flags
i have a project that uses cc in it's deep backend, and i need to remove one of the flags added deep inside (which i can't remove). i added this function (it has some shit time complexity, but it's like 2am in my country, so i don't care). so yeah, i made this function to fix it and remove this function