Remove some dead code in objects_from_file
Since https://github.com/rust-lang/cc-rs/pull/684 was merged, it is
impossible for `obj` to not start with dst - if `obj` is absolute or
contains `../` in its Path, then the path will be hashed, and the file
will still be placed under dst.

As such, this obj.starts_with(&dst) check is always true.