Use RUSTC_WRAPPER if no other wrapper is provided
Previously, the `rustc_wrapper_fallback` was only called if the tool
 was **not** a full path.
With this patch, the `RUSTC_WRAPPER` is always used, even if the tool
provided is an exact path.
Providing a wrapper manually is still possible and overrides the
`RUSTC_WRAPPER`.

If the path to the tool includes spaces it is otherwise impossible to
provide a compiler cache like sccache through the CXX or CC environment
variables, as the arguments will be split up incorrectly.

See also: https://github.com/corrosion-rs/corrosion/pull/474