Add empty `jobserver` feature.
Since we removed implicit features, anyone that used this crate
and specifically enabled `jobserver` would now be broken as this
crate no longer had that feature (or a couple of others).

Add a placeholder `jobserver` feature that doesn't enable `parallel`
(just as it didn't enable it before).

The other features removed weren't likely to be enabled by anyone.