Solarish-ish: define _REENTRANT by default
Descendants of Solaris define some C APIs in a manner that results in
incorrect behavior when used from code with mixed multi-threading
support. One example here is that illumos defines errno such that errno
used in a process' first thread happens to be the same errno as global
errno; in https://github.com/oxidecomputer/helios/issues/192 this
results in a Tokio panic as it appears that `read()` returned with errno
of 0!

Given the prevalence of multi-threaded programs these days, it would be
nice for the compilation environment to not need additional flags
"basically always." But until illumos' headers are adjusted in this
direction, `cc` can at least define `_REENTRANT` as a default for these
systems.