Use `$WASI_SDK_PATH` on WASI targets by default
This commit updates compiler detection logic to use the
`$WASI_SDK_PATH` variable, if set, by default when looking up a compiler
for WASI targets. This is commonly set in installations of wasi-sdk and
it can be helpful to avoid the need to further configure how to find a
WASI-specific compiler.

CI for WASI was expanded here from `wasm32-wasip1-threads` to other WASI
targets as well to ensure it works on all of Rust's targets.