Fix building for Mac Catalyst (#1577)
Without this change, we end up passing `aarch64-apple-ios26.0-macabimacabi` to the compiler, which is definitely wrong.

The core change here is to handle this in `from_cargo_environment_variables` as well, the rest is just a cleanup now that `target.env` contains the canonical value (and `target.abi`).

See also the previous #1517 and #1534.