Stop passing an invalid target to `llvm-mingw`'s cross-compilation wrappers (#1495)
The cross-compilation toolchains distributed by the `llvm-mingw` project
for building code for the `*-pc-windows-gnullvm` targets have a wrapper
script in place of their `clang` binaries that calls them with a
specific target, much like it has been the case with the Android NDK.
When `cc-rs` sets a explicit `--target` flag, the wrapper script breaks
down, and is no longer able to use the proper linker and libraries.

To improve on this situation and make it possible to cross-compile Rust
projects that depend on building C code with `cc-rs`, let's skip passing
this flag when cross-compilation with such a wrapper script is
attempted.